+++ .vscode/settings.json
@@ -3,6 +3,7 @@
 		"Anticheat",
 		"appguid",
 		"appname",
+		"Arkham",
 		"catcache",
 		"chrono",
 		"compatdata",
@@ -13,6 +14,7 @@
 		"HKEY",
 		"ICONERROR",
 		"IDYES",
+		"Infobox",
 		"installsource",
 		"isfreeapp",
 		"KTJZNR",
